+++ crates/guest-write-file/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.1.68](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.67...guest-write-file-v0.1.68) (2026-08-14)
+
 ## [0.1.67](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.66...guest-write-file-v0.1.67) (2026-08-14)
 
 ## [0.1.66](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.65...guest-write-file-v0.1.66) (2026-08-13)